+++ public/admin/_app/immutable/nodes/24.Cg65erGM.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/CrfwAsjM.js";export{e as component};