+++ public/admin/_app/immutable/nodes/41.IwfVHcTy.js
@@ -1,0 +1,1 @@
+import{It as e}from"../chunks/DFnvYPbU.js";import{c as t}from"../chunks/CDZhc2eX.js";import{t as n}from"../chunks/CApOcFjb.js";var r=e({load:()=>i}),i=()=>{throw n(308,`${t}/settings/users`)};export{r as universal};